requirements.txt
# Core dependencies
python>=3.8
pycryptodome>=3.14.1
cryptography>=36.0.0

# Cloud APIs
google-api-python-client>=2.48.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6
dropbox>=11.32.0
msal>=1.17.0
requests>=2.27.0

# Utilities
click>=8.0.0
colorama>=0.4.4
tqdm>=4.62.0
python-dotenv>=0.19.0
pyyaml>=6.0
python-magic>=0.4.24

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.7.0
pytest-benchmark>=3.4.1
hypothesis>=6.39.3

# Documentation
sphinx>=4.4.0
sphinx-rtd-theme>=1.0.0
myst-parser>=0.17.0

# Development
black>=22.1.0
flake8>=4.0.0
mypy>=0.940
isort>=5.10.0
pre-commit>=2.17.0

# Performance
numpy>=1.22.0
pandas>=1.4.0
matplotlib>=3.5.0

# Security
bandit>=1.7.0
safety>=1.10.3
